Move the call to drop() into release()
This causes us to decrement all of the references
to our children during release() which is what the pseudocode
in the Bacon Rajan paper does. Removing drop() from free()
also avoids the droping during MarkRoots() which can cause
problems with the reference graph as suggested in #18.

This is an alternative solution to the problem addressed by
#29