Fix a Miri error during test_retain_weak.
We were getting an error like:

Undefined Behavior: not granting access to tag <369452> because that would
remove [Unique for <373036>] which is strongly protected because it is an
argument of call 153932

This was caused by us accessing our own data() to check the reference counts
on a self cyclic reference during a call to `drop()` that has a mut
ref.

We fix it by only calling drop() on the value in the CcBoxPtr instead of
on the CcBoxPtr itself. This makes it so the mut reference doesn't cover
the CcBoxData. It comes at the cost of needing to an extra indirect call
to get a trait object that covers just the value because we don't have a
way to call trait object methods with just a raw pointer.